requirements.txt
aiogram
beautifulsoup4
requests
lxml
bs4
python-dotenv
